refactor/data-fetching-logic
